requirements.txt
fastapi
uvicorn[standard]
pydantic
httpx
python-dotenv
replicate
moviepy
requests